Move conductor out to root directory
